feat: create config file if not exists
